+++ apidump/net.go
@@ -6,6 +6,8 @@ import (
 	"os"
 	"strconv"
 	"strings"
+	"sync"
+	"time"
 
 	"github.com/akitasoftware/akita-libs/api_schema"
 	"github.com/akitasoftware/go-utils/optionals"
@@ -13,7 +15,9 @@ import (
 	"github.com/postmanlabs/postman-insights-agent/architecture"
 	"github.com/postmanlabs/postman-insights-agent/consts"
 	"github.com/postmanlabs/postman-insights-agent/env"
+	"github.com/postmanlabs/postman-insights-agent/pcap"
 	"github.com/postmanlabs/postman-insights-agent/printer"
+	"github.com/postmanlabs/postman-insights-agent/telemetry"
 )
 
 // An interface that's compatible with net.Interface so we can use mock
@@ -177,6 +181,40 @@ type pcapPermErr struct {
 
 func (pe pcapPermErr) Error() string {
 	return fmt.Sprintf("failed to read packets from interface %s: %v", pe.iface, pe.err)
+}
+
+// Check if we have permission to capture packets on the given set of
+// interfaces.
+func checkPcapPermissions(interfaces map[string]interfaceInfo, targetNetworkNamespaceOpt optionals.Optional[string]) map[string]error {
+	printer.Debugf("Checking pcap permissions...\n")
+	start := time.Now()
+
+	var wg sync.WaitGroup
+	wg.Add(len(interfaces))
+	errChan := make(chan *pcapPermErr, len(interfaces)) // buffered enough to never block
+	for iface := range interfaces {
+		go func(iface string) {
+			defer wg.Done()
+			h, err := pcap.GetPcapHandle(iface, 1600, true, pcap.BlockForever, targetNetworkNamespaceOpt)
+			if err != nil {
+				telemetry.Error("pcap permissions", err)
+				errChan <- &pcapPermErr{iface: iface, err: err}
+				return
+			}
+			h.Close()
+		}(iface)
+	}
+
+	wg.Wait()
+	printer.Debugf("Check pcap permission done after %s\n", time.Since(start))
+	close(errChan)
+	errs := map[string]error{}
+	for pe := range errChan {
+		if pe != nil {
+			errs[pe.iface] = pe
+		}
+	}
+	return errs
 }
 
 // Returns BPF filter for inbound API spec traffic on each interface.
