+++ pcap/pcap.go
@@ -2,10 +2,12 @@ package pcap
 
 import (
 	"net"
+	"time"
 
 	"github.com/akitasoftware/go-utils/optionals"
 	"github.com/google/gopacket"
 	_ "github.com/google/gopacket/layers"
+	"github.com/google/gopacket/pcap"
 	"github.com/pkg/errors"
 	"github.com/postmanlabs/postman-insights-agent/printer"
 )
@@ -13,6 +15,7 @@ import (
 const (
 	// The same default as tcpdump.
 	defaultSnapLen = 262144
+	BlockForever   = pcap.BlockForever
 )
 
 type pcapWrapper interface {
@@ -21,6 +24,59 @@ type pcapWrapper interface {
 }
 
 type pcapImpl struct{}
+
+func (p *pcapImpl) capturePackets(done <-chan struct{}, interfaceName, bpfFilter string, targetNetworkNamespaceOpt optionals.Optional[string]) (<-chan gopacket.Packet, error) {
+	handle, err := GetPcapHandle(interfaceName, defaultSnapLen, true, BlockForever, targetNetworkNamespaceOpt)
+	if err != nil {
+		return nil, err
+	}
+
+	if bpfFilter != "" {
+		if err := handle.SetBPFFilter(bpfFilter); err != nil {
+			handle.Close()
+			return nil, errors.Wrap(err, "failed to set BPF filter")
+		}
+	}
+
+	// Creating the packet source takes some time - do it here so the caller can
+	// be confident that pakcets are being watched after this function returns.
+	packetSource := gopacket.NewPacketSource(handle, handle.LinkType())
+	pktChan := packetSource.Packets()
+
+	// TODO: tune the packet channel buffer
+	wrappedChan := make(chan gopacket.Packet, 10)
+	go func() {
+		// Closing the handle can take a long time, so we close wrappedChan first to
+		// allow the packet consumer to advance with its processing logic while we
+		// wait for the handle to close in this goroutine.
+		defer func() {
+			close(wrappedChan)
+			handle.Close()
+		}()
+
+		startTime := time.Now()
+		count := 0
+		for {
+			select {
+			case <-done:
+				return
+			case pkt, ok := <-pktChan:
+				if ok {
+					wrappedChan <- pkt
+
+					if count == 0 {
+						ttfp := time.Since(startTime)
+						printer.Debugf("Time to first packet on %s: %s\n", interfaceName, ttfp)
+					}
+					count += 1
+				} else {
+					return
+				}
+			}
+		}
+	}()
+	return wrappedChan, nil
+}
 
 func (p *pcapImpl) getInterfaceAddrs(interfaceName string) ([]net.IP, error) {
 	iface, err := net.InterfaceByName(interfaceName)
